backend/requirements.txt
asgiref==3.8.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.10
sqlparse==0.5.1
tzdata==2024.1
